java-cdk: split DestinationAcceptanceTest record-oriented methods to new classes
